Remove old module and benchmarks
