+++ packages/css/test.js
@@ -1,6 +1,6 @@
 import test from 'ava'
 import { width, space } from 'styled-system'
-import css, { createCSS } from './alt'
+import css from './index'
 
 const theme = {
   colors: {
@@ -96,16 +96,6 @@ test('returns nested responsive styles', t => {
   })
 })
 
-// breaking change
-test.skip('props override default styles', t => {
-  const res = css({
-    color: 'primary',
-  })({ theme, color: 'secondary' })
-  t.deepEqual(res, {
-    color: 'cyan'
-  })
-})
-
 test('handles all core styled system props', t => {
   const res = css({
     m: 0,
@@ -152,20 +142,16 @@ test('works with the css prop', t => {
 
 test.todo('handles custom transforms')
 
-test.skip('createCSS returns a custom css function', t => {
-  const func = createCSS([ width, space ])
-  const res = func({
-    width: [ 1/2, 1/4 ],
-    m: 3,
-    color: 'primary'
-  })({ theme })
-  t.is(typeof func, 'function')
-  t.deepEqual(res, {
-    width: '50%',
-    '@media screen and (min-width: 40em)': {
-      width: '25%',
-    },
-    margin: 16,
+// breaking
+// test.skip('createCSS returns a custom css function')
+
+// breaking
+test.skip('props override default styles', t => {
+  const res = css({
     color: 'primary',
+  })({ theme, color: 'secondary' })
+  t.deepEqual(res, {
+    color: 'cyan'
   })
 })
+
